Refactored to move out code from autogenerated files as per PR review
