+++ lib/CMakeLists.txt
@@ -67,6 +67,7 @@ if(LIBOCPP_ENABLE_V201)
             ocpp/v201/init_device_model_db.cpp
             ocpp/v201/notify_report_requests_splitter.cpp
             ocpp/v201/message_queue.cpp
+            ocpp/v201/profile.cpp
             ocpp/v201/ocpp_types.cpp
             ocpp/v201/ocsp_updater.cpp
             ocpp/v201/monitoring_updater.cpp
