+++ tests/lib/ocpp/v201/test_composite_schedule.cpp
@@ -8,6 +8,7 @@
 #include "ocpp/v201/device_model_storage_sqlite.hpp"
 #include "ocpp/v201/init_device_model_db.hpp"
 #include "ocpp/v201/ocpp_types.hpp"
+#include "ocpp/v201/profile.hpp"
 #include "ocpp/v201/utils.hpp"
 #include <algorithm>
 #include <boost/uuid/uuid_generators.hpp>
